viber: wip - replace bundled dependencies
The bundled Qt6 dependencies perform hard-coded lookups to FSH paths.
These are hard to patch directly although it may be possible by using
symlinks and keeping the paths sort. Nixpkgs patch the Qt6 source
so that this is more controllable.

This commit attempts to replace the bundled libraries and plugins with
nixpkgs ones. In order to do so, we need to symlink to the installation
directory for Qt6 to properly pick them up.

The commit does not currently work, Viber starts properly but nothing is
displayed. A binding loop is reported which may be atttributed to the
different Qt6 version semantics regarding resizing.